Moves templ-multi to lib
