+++ lib/templ-multi.js
